requirements.txt
xgboost>=2.0
lightgbm>=4.0
catboost>=1.2
scikit-learn>=1.3
pandas>=2.0
numpy>=1.24
requests>=2.31
beautifulsoup4>=4.12
lxml>=4.9
python-telegram-bot>=20.0
python-dotenv>=1.0
apscheduler>=3.10
joblib>=1.3
